Adding check whether an unsupported entitlement is used, that we can not add on the API.
- We check this before checking if App ID exist and is valid as as far we know,
these needed to be added in the provisioning profile generation phase.
So even if the user creates App ID manually, the step can not generate the correct profiles.
The extra entitlements require Apple approval, which we can not test,
so choose to fail early and make the user use the Certificates Installer Step.